backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
audioread==3.0.1
backports.tarfile==1.2.0
boltons==25.0.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
contourpy<=1.3.1
crashtest==0.4.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
distlib==0.3.9
docopt==0.6.2
dulwich==0.22.8
environs==14.1.1
exceptiongroup==1.2.2
# Changed FastAPI version to be compatible with spotdl
fastapi==0.103.2
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fonttools==4.56.0
getmac==0.9.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
joblib==1.4.2
jsonpickle==4.0.5
keyring==25.6.0
kiwisolver<=1.4.8
lazy_loader==0.4
librosa<=0.11.0
llvmlite<=0.44.0
marshmallow<=3.26.1
matplotlib<=3.10.0
mido<=1.3.3
more-itertools<=10.6.0
msgpack<=1.1.0
music21<=9.5.0
noisereduce<=3.0.3
ntplib<=0.4.0
numba<=0.61.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pbs-installer==2025.3.17
pedalboard==0.9.16
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
pooch==1.8.2
pretty_midi==0.2.10
pyarrow<=19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
RapidFuzz==3.12.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy<=1.15.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.0
sourcedefender==15.0.1
soxr==0.5.0.post1
# Changed Starlette version to be compatible with FastAPI 0.103.2
starlette==0.27.0
TgCrypto-pyrofork==1.2.7
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.3.19.19
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.23.2
virtualenv==20.29.3
webcolors==24.11.1
xattr==1.1.4
zipp==3.21.0
zstandard==0.23.0
yt-dlp==2025.3.31
ffmpeg-python==0.2.0
spotdl==4.2.11
spotipy==2.25.1
sqlmodel==0.0.24
psycopg2==2.9.10
python-jose[cryptography]==3.4.0
email-validator==2.2.0
dnspython==2.7.0
python-dotenv==1.1.0